Move consumer_group related test into a separate file.
